requirements.txt
pillow~=11.0.0
torchvision~=0.20.1+cu124
tqdm~=4.67.1
pandas~=2.2.3
transformers~=4.47.0
nltk~=3.9.1
torch~=2.5.1+cu124
numpy~=1.26.3
scikit-learn~=1.6.0
streamlit~=1.41.1
sentence-transformers~=3.3.1
clip~=1.0
Flask~=3.1.0
faiss-cpu~=1.9.0.post1